Revert "Enhance middleware authentication flow and improve user experience"
