+++ lib/supabase.ts
@@ -304,63 +304,45 @@ export async function refreshSession() {
   }
 }
 
-// Enhanced function for more robust user retrieval that tries multiple approaches
+// A more robust way to get the user with multiple fallbacks
 export async function robustGetUser(): Promise<any | null> {
   if (!supabase) {
-    console.error("robustGetUser: Supabase client not initialized");
+    console.warn("Supabase client not initialized");
     return null;
   }
 
   try {
-    console.log("robustGetUser: Starting robust user retrieval");
-    
-    // First, try to get session directly - this is the fastest method
+    // Try with getSession first
     const { data: sessionData, error: sessionError } = await supabase.auth.getSession();
     
     if (sessionError) {
-      console.error("robustGetUser: Error getting session:", sessionError);
+      console.warn("Error getting session:", sessionError.message);
     } else if (sessionData?.session?.user) {
-      console.log("robustGetUser: Found user via getSession:", sessionData.session.user.id);
       return sessionData.session.user;
-    } else {
-      console.log("robustGetUser: No session found via getSession");
     }
     
-    // If we don't have a user yet, try refreshing the session
-    console.log("robustGetUser: Attempting to refresh the session");
-    try {
-      const { data: refreshData, error: refreshError } = await supabase.auth.refreshSession();
-      
-      if (refreshError) {
-        console.error("robustGetUser: Error refreshing session:", refreshError);
-      } else if (refreshData?.session?.user) {
-        console.log("robustGetUser: Found user via refreshSession:", refreshData.session.user.id);
-        return refreshData.session.user;
-      } else {
-        console.log("robustGetUser: No user found after session refresh");
-      }
-    } catch (refreshCatchError) {
-      console.error("robustGetUser: Exception during session refresh:", refreshCatchError);
-    }
-    
-    // Last attempt: Try getUser directly
-    console.log("robustGetUser: Final attempt using getUser");
+    // If that fails, try with getUser
     const { data: userData, error: userError } = await supabase.auth.getUser();
     
     if (userError) {
-      console.error("robustGetUser: Error getting user:", userError);
-      return null;
-    }
-    
-    if (userData?.user) {
-      console.log("robustGetUser: Found user via getUser:", userData.user.id);
+      console.warn("Error getting user:", userError.message);
+    } else if (userData?.user) {
       return userData.user;
     }
     
-    console.log("robustGetUser: No user found after all attempts");
+    // As a last resort, try to refresh the session
+    const { data: refreshData, error: refreshError } = await supabase.auth.refreshSession();
+    
+    if (refreshError) {
+      console.warn("Error refreshing session:", refreshError.message);
+    } else if (refreshData?.session?.user) {
+      return refreshData.session.user;
+    }
+    
+    console.warn("No user found after all attempts");
     return null;
   } catch (error) {
-    console.error("robustGetUser: Unexpected error:", error);
+    console.error("Error in robustGetUser:", error);
     return null;
   }
 }
